requirements.txt
discord.py>=2.3.0
apscheduler>=3.10.0
pytz>=2023.3